upd
git-svn-id: file:///Users/demichel/SVNREPOS/CODE@4931 fea3b138-1bf5-0310-9da5-cbb00e418fd0

+++ MDsimul++/hard_ellipses/calc_veff_hardell.cpp
@@ -37,7 +37,7 @@ int main (int argc, char **argv)
       exit(1);
     }
   if (argc >= 2)
-    ttmax=atoi(argv[1]);
+    ttmax=atoll(argv[1]);
   else
     ttmax = 1000000;
